pyproject.toml
[project]
name = "telegram_bot"
version = "0.1.0"
description = "Telegram Bot"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "python-telegram-bot[ext]>=21.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.4.0",
    "vosk==0.3.45",
    "soundfile==0.12.1",
    "ffmpeg-python==0.2.0",
    "pydub==0.25.1",
    "langchain-gigachat>=0.1.5",
    "langchain-ollama>=0.3.3",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.4.3",
    "watchdog>=3.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"] 